requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
mysqlclient==2.2.1
Werkzeug==3.0.0
PyMySQL==1.1.0
alembic==1.13.1
python-dotenv==1.0.0
flask_wtf==3.2.1
